add history controls to chat, add analysis button to history
