+++ Cargo.toml
@@ -10,12 +10,12 @@ opt-level = 'z'
 
 [workspace.dependencies]
 leptos = { version = "0.6" , features = ["nightly"] }
-leptos-use = { version = "0.11" }
+leptos-use = { version = "0.12" }
 leptos_meta =  { version = "0.6" , features = ["nightly"] }
 leptos_router =  { version = "0.6" , features = ["nightly"] }
 leptos_actix =  { version = "0.6" }
 leptos_icons = { version = "0.3"}
-icondata = {version = "0.3"}
+icondata = {version = "0.4"}
 actix-web = { version = "4",  features = ["macros"] }
 actix-files = { version = "0.6" }
 actix-web-actors = { version  = "4.3.0" }
@@ -40,9 +40,9 @@ simple_logger = "5.0"
 thiserror = "1"
 anyhow = "1"
 tokio = { version = "1.39.2", features = ["full"] }
-wasm-bindgen = "0.2.92"
+wasm-bindgen = "0.2.93"
 wasm-bindgen-futures = "0.4"
-web-sys = {version = "0.3.69", features = ["AbortController", "AbortSignal", "AudioContext", "AudioBuffer", "AudioBufferSourceNode", "AudioDestinationNode", "Blob", "Clipboard", "HtmlDocument", "SvgPoint", "SvgsvgElement", "SvgGraphicsElement", "SvgRect", "SvgMatrix", "Url", "Window"] }
+web-sys = {version = "0.3.70", features = ["AbortController", "AbortSignal", "AudioContext", "AudioBuffer", "AudioBufferSourceNode", "AudioDestinationNode", "Blob", "Clipboard", "HtmlDocument", "SvgPoint", "SvgsvgElement", "SvgGraphicsElement", "SvgRect", "SvgMatrix", "Url", "Window"] }
 bb8 = { version = "0.8" }
 diesel = { version = "2.2", features = ["postgres", "chrono", "uuid", "serde_json"] }
 diesel-async = { version = "0.5", features = ["postgres", "bb8"] }
